+++ testing/matrix-sdk-test/src/sync_builder/invited_room.rs
