+++ testing/matrix-sdk-test/src/sync_builder/joined_room.rs
