+++ testing/matrix-sdk-test/src/sync_builder/left_room.rs
